+++ src/Gaurav.Danani.WooliesX.Application/Trolleys/Queries/GetTrolleyTotal/GetTrolleyTotalQuery.cs
@@ -2,7 +2,7 @@
 
 namespace Gaurav.Danani.WooliesX.Application.Trolleys.Queries.GetTrolleyTotal
 {
-    public class GetTrolleyTotalQuery : IRequest<double>
+    public class GetTrolleyTotalQuery : IRequest<decimal>
     {
         public TrolleyTotalRequest Trolley { get; set; }
     }
